Task 12/06
- Automatically create a combo whenever a hotel is created

- Validate range date when creating new peak day record

- Add icon to combo list in customer-client side

- Prevent admin/employee from deleting related record

- Remove duration field in backup table

+++ server/src/controllers/combo.js
@@ -1,5 +1,6 @@
 import mongoose from "mongoose";
 import Combo from "../models/combo.js";
+import Booking from "../models/booking.js";
 import Log from "../models/log.js";
 import { STRING, INTEGER } from "../constants/constants.js";
 
@@ -78,6 +79,14 @@ export const deleteCombo = async (req, res) => {
     return res.status(404).send("No combo with that id");
   }
   try {
+    // CHECK RELATED RECORD
+    let delete_id = mongoose.Types.ObjectId(id);
+    const related_booking = await Booking.findOne({ combo_list: delete_id });
+    if (related_booking) {
+      return res.status(409).send(STRING.DELETE_RELATED_RECORD);
+    }
+
+    // PROCESS
     const TIME_STAMP = new Date();
     await Combo.findOneAndRemove({ _id: id });
     await logAction(req._id, INTEGER.LOG_DELETE, TIME_STAMP);
